Add EFI and GPT support

+++ include/aplus/utils/ringbuffer.h
@@ -50,12 +50,14 @@ __BEGIN_DECLS
 void ringbuffer_init(ringbuffer_t* rb, size_t size);
 void ringbuffer_destroy(ringbuffer_t* rb);
 void ringbuffer_reset(ringbuffer_t* rb);
+
 int ringbuffer_is_full(ringbuffer_t* rb);
 int ringbuffer_is_empty(ringbuffer_t* rb);
 size_t ringbuffer_available(ringbuffer_t* rb);
 size_t ringbuffer_writeable(ringbuffer_t* rb);
-int ringbuffer_write(ringbuffer_t* rb, const void* buf, size_t size);
-int ringbuffer_read(ringbuffer_t* rb, void* buf, size_t size);
+
+ssize_t ringbuffer_write(ringbuffer_t* rb, const void* buf, size_t size);
+ssize_t ringbuffer_read(ringbuffer_t* rb, void* buf, size_t size);
 
 __END_DECLS
